Move SecureSettingsRepository to shared.
We need a way to access notifications-related secure settings in new
sysui recommended architecture code. WallpaperPicker already had code
for this (in the form of a NotificationsInteractor,
NotificationsRepository and SecureSettingsRepository), so instead of
having multiple sources of truth for the same type of data, I moved
these to the shared package so both SystemUI and WallpaperPicker can
access them (see other CLs in topic).

Also added a provider for the background scope, as it's needed in
another CL in this topic.

Bug: 293167744
Test: manual + presubmit
Flag: NONE
Change-Id: I288b7bac4122911c5662751398970c5ff3e7d220