Preview wallpaper with crop hints in WallpaperPreviewActivity
Ensure that the current multi-crop wallpaper is previewed with correct
cropping by making sure that crop hints gets piped to the
WallpaperPreviewActivity.

This involves updating WallpaperMetadata, WallpaperInfo,
StaticWallpaperData etc. to include crop hints with the updated type. It
also involves making sure that the correct wallpaper info is received
in the entry point to WallpaperPreviewActivity in
ScreenPreviewSectionController; to do so, DefaultWallpaperRefresher is
updated to get the current crop hints correctly. Then, when converting
from WallpaperInfo to WallpaperModel, crop hints needs to be passed
along, so changes were made in the WallpaperModelFactory classes.
Lastly, WallpaperPreviewActivity sets the crophints in the
WallpaperPreviewViewModel.

Flag: ACONFIG com.android.wallpaper.multi_crop_preview_ui_flag DEVELOPMENT
Test: manually verified
Bug: 303317694
Bug: 322220597
Change-Id: I89fabb4d43d742092280cb5257b027a463305533