Live wallpaper stretch fix for foldables
The root cause for this issue is that we only create 1 engine and mirror
the live wallpaper preview to all 4 surfaces. But 2 of them are of
differet display size.

In the fix we create 2 engines for general live wallpapers. So that one
engine is for unfolded and one for folded.

However, the legacy live wallpapers do not support 2 engines. We
fallback to single engine and make sure that we mirror with the
correspondent display size to have the correct scale.

Test: Manually tested on all kinds of live wallpapers
Bug: 328484124
Flag: ACONFIG com.android.wallpaper.multi_crop_preview_ui_flag TEAMFOOD
Change-Id: Id5ea1250a16aafc080b1fe9ebd99a49f38a41740