Fix Multi-Crop Set Wallpaper Transition edge case
With the "Multi-Crop Set Wallpaper Transition" change applied,
when navigating from small to full preview then quickly returning to
small preview, previews disappear. This is because when the transition
from small to full preview starts, we set preview surface views to be
invisible, but in this edge case, the small preview does not go through
destroy, and as a result does not go through create or onCreateView
either, simply starting and resuming, causing the views to remain
invisible. To fix this, make sure views are visible on lifecycle start.

Flag: ACONFIG com.android.wallpaper.multi_crop_preview_ui_flag TEAMFOOD
Bug: 325035174
Bug: 303318205
Test: manually verified on handheld and foldable
Change-Id: I832f651307fefe82e182d52c15800fa54d87a79a