Take into account current wallpaper effects for double engine preview
Add an "isEffectsWallpaper" to LiveWallpaperData based on the component
and use that to determine if we need to use single engine or double for
preview.
Remove "effectNames" from DownloadableWallpaperData as it doesn't make
sense there.
Skip the effects flow if not available on the device.

Bug: 330141698
Test: manually verify with different types of live and effects wallpaper
as current
Test: atest WallpaperPickerTests
Flag: ACONFIG com.android.wallpaper.multi_crop_preview_ui_flag TEAMFOOD

Change-Id: I91ad46e1d13e8ca2ae8236cecadfa9d84b17aefa