Fix using wrong lockscreen wallpaper
When static lock wallpaper exist and getWallpaperFile(LOCK) is null, the
lock wallpaper is a static default wallpaper, in this case use
DefaultWallpaperInfo that has BuiltInWallpaperAsset whereas
CurrentWallpaperInfo relies on getWallpaperFile that is null.

Reload main screen preview in this case so it's showing the correct
wallpaper preview.

Also fix a case where suspendCancellableCoroutine is not resuming when
decodeRawDimensions returns null.

Flag: ACONFIG com.android.window.flags.multi_crop TEAMFOOD
Bug: 331557838
Test: set home wallpaper only
Change-Id: I2af0ca00ede6f112813498c129409ed209d185e7