Fix live wallpaper download & set issue
After live wallpaper download, the preview is not updated to display a
live wallpaper, because collecting from the wallpaper model flow is
blocked in the SmallWallpaperPreviewBinder by binding the
StaticWallpaperPreviewBinder. The static wallpaper bind function is
suspended until view lifecycle is stopped due to its repeat on lifecycle
started function call. To adapt to the parent's job lifecycle, which is
created in the SurfaceHolder surfaceCreated callback and cancelled in
the surfaceDestroyed callback, we launch coroutines from the parent's
coroutine scope and remove repeat on lifecycle started altogether, to
ensure it does not block parent flow collection, and to ensure that it
gets destroyed along with the parent coroutine in the surfaceDestroyed
callback.

Flag: ACONFIG com.android.wallpaper.multi_crop_preview_ui_flag TEAMFOOD
Bug: 334862324
Bug: 333819771
Bug: 330379641
Test: manually verified by downloading & setting live wallpaper
Test: manually verified static wallpaper previews & sets correctly
Test: used print logs to test proper coroutine cancellation, see steps
in b/330379641#comment7

Change-Id: Ida36cb1e9950a7dd0f5cc68ab30fe2ffe423a3ae