Add test cases to DisplayUtilsTest and refine FakeDisplaysProvider
Changed setDisplays in FakeDisplaysProvider to take in
FakeDisplayConfig, a new data class that contains possible
customizations to a fake display.

Added tests to functions that are testable and test-worthy. Any tests
that rely on isOnWallpaperDisplay is not tested at the moment due to
the inability to set uniqueId on displays created with
ShadowDisplayManager, but it could be worth it to create a fake version
of the function for dependent functions to be tested in the future.

Flag: NONE
Test: DisplayUtilsTest
Bug: 335647826
Change-Id: Id5d6bc94fcf047a7e7a19f572aa3f912270373e9