Fix screen preview talkback
Create separate lock and home screen preview layouts for better
accessibility, so only lock preview has click interceptor and clock
carousel. Allow talkback to handle the clock carousel separately from
the preview, so that both lock and home previews are focasable. Add
onClick as an input to ScreenPreviewBinder, so that when it is not set,
like in color picker fragment, the preview is not clickable, and doesn't
announce "double tap to activate", and when it is set in the home and
lock screen preview sections, the preview is clickable, and announces
"double tap to activate".

Flag: NONE
Fixes: 320353328
Test: manually verified with talkback on and off
Change-Id: I2abedd24fad3c23b536bdc398987d8a33067ea40