Revert^2 "Fix screen preview talkback"
This reverts commit 1bb90039ffce00928d9109e01b60d61af11a83cc.

Reason for revert: bringing original change back ag/27195877 with fix
for b/338667501, in which Wallpaper Picker crashes due to a layout
issue. The layout issue is caused by switching the original screen
preview section layout for a simplified version on the home screen,
with ScreenPreviewClickView removed. This causes a crash in certain
devices when entering the if check in ScreenPreviewSectionController for
isTwoPaneAndSmallWidth. In this CL, I've removed the layout change,
which surprisingly doesn't impact the new talkback behavior.

Change-Id: Ic888227398645ceb239cfa0f7b29f95de5fca4a8
Flag: NONE
Fixes: 320353328
Test: manually verified with talkback on and off
Test: manually verified on handheld and foldable