Use correct wallpaper model after rotation
When onCreated the activity sets the wallpaper model from the intent to
the repository. For downloaded wallpaper the model will be live
wallpaper in the repository  but the one came with intent is still static.
To fix this, if the activity created after a config change we don't set
the wallpaper model again, so the repository can use the right model.

Flag: ACONFIG com.android.window.flags.multi_crop TRUNKFOOD
Fixes: 336425210
Test: download wallpaper then rotate
Change-Id: I5ecf6f363a88eb5ac45d88cb4a4c5022a92bd6a1