Fix number of columns in category selector talkback
The default column span size of 2 is a legacy to enable placing 2
categories in a row or 3 categories in a row. This is no longer relevant
in the creative wallpaper enabled category picker since there are no
cases of 2 categories in a row. This fixes the talkback issues of
announcing there are 6 columns when visually there are only 3.

Flag: NONE
Bug: 322735050
Test: manually verified using talkback
Change-Id: I5279b760edfbae22e21dbf494f3372a084ec168a